test(deliverable-template): test_dag.py 530行をクラス単位で3ファイルに分割 (Issue #122)
Rams 再レビュー指摘対応: test_dag.py (530行) が500行を超えているため
責務単位で3ファイルに分割した。

- test_ref_validation.py  (101行): TC-IT-DTH-002/003
  ref不在・自己参照バリデーション（入力バリデーション責務）
- test_cycle_detection.py (310行): TC-IT-DTH-004/005
  推移的循環・深度制限（循環検出責務）
- test_dag_normal.py      (158行): TC-IT-DTH-022
  合法 diamond DAG 正常系（合法 DAG 受容責務）
